🖊️Refactored Save ans SaveProcess Added Priority files and large file locking (to be tested) Fixed MVVM by removind model commands in Save

+++ EasySave/Models/StateLogger.cs
@@ -86,7 +86,7 @@ namespace EasySave.Models
                                         {
                                             save.Transfering = true;
                                             save.FilesRemaining = element.GetProperty("filesRemaining").GetInt32();
-                                            save.SizeRemaining = element.GetProperty("sizeRemaining").GetInt32();
+                                            save.SizeRemaining = element.GetProperty("sizeRemaining").GetInt64();
                                             save.CurrentSource = element.GetProperty("currentSource").GetString() ?? "";
                                             save.CurrentDestination = element.GetProperty("currentDestination").GetString() ?? "";
                                         }
